03-computer-vision/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
polars==1.33.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.196
ultralytics-thop==2.0.17
urllib3==2.5.0
